// src/types.ts
export type SUStatus =
  | "primary"
  | "secondary"
  | "success"
  | "info"
  | "warning"
  | "failure"
  | "debug";

export type SUOrientation = "vertical" | "horizontal";

export type SUOutline = "invisible" | "solid" | "dashed";

export type SUPresentation = "block" | "line";

export type SUTrigger = "hover" | "click";